feat(phase5): wire targetDate through discovery interface
- Update ChainAdapter.Discover() to accept targetDate parameter
- Update BaseChainAdapter and KonzumAdapter implementations
- Remove SetDiscoveryDate method (no longer needed)
- Pass targetDate from DiscoverPhase to adapter.Discover()
- Add registry for chain adapter management
- Add pipeline phases (discover, fetch, parse, normalize)
- Add ingest HTTP handler for triggering ingestion
- Update main.go with internal admin routes
- Add .env support to config loading

+++ services/price-service/internal/http/ratelimit/retry.go
@@ -67,7 +67,7 @@ func CalculateRateLimitBackoff(attempt int, config Config, retryAfterHeader *str
 }
 
 // Sleep blocks for the specified duration in milliseconds
-func Sleep(ms int) {
+func Sleep(ms int64) {
 	time.Sleep(time.Duration(ms) * time.Millisecond)
 }
